backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.2
anthropic==0.52.0
pymupdf==1.25.5
python-docx==1.1.2
aiosqlite==0.21.0
python-multipart==0.0.20
pydantic==2.11.3
